// src/api/generated/models/accessPolicyResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */

export interface AccessPolicyResponse {
  roleCode?: string;
  resourceType?: string;
  action?: string;
  scope?: string;
  validFrom?: string;
  validTo?: string;
  current?: boolean;
}